change variable type to allow more integers for ids (#126)

+++ src/rate_fns/rate_fn_storage.rs
@@ -73,7 +73,7 @@ pub fn load_rate_fns(context: &mut Context) -> Result<(), IxaError> {
 
 #[derive(Deserialize)]
 pub struct EmpiricalRateFnRecord {
-    id: u8,
+    id: u32,
     time: f64,
     value: f64,
 }
